pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestration-pipelines"
dynamic = ["version"]
authors = [
  { name="Google LLC", email="googleapis-packages@google.com" },
]
description = "Orchestration Pipelines"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Framework :: Apache Airflow",
]
dependencies = [
    "pytz",
    "PyYAML",
    "apache-airflow>=2.9.0",
    "google-cloud-storage",
    "dbt-core>=1.0.0",
    "protobuf>=4.25.0",
    "apache-airflow-client",
    "google-auth",
    "google-cloud-dataproc>=5.25.0",
    "google-cloud-dataform>=0.9.0",
    "apache-airflow-providers-google>=19.0.0",
    "tenacity",
]

[project.license]
text = "Apache-2.0"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestration_pipelines_lib*", "orchestration_pipelines_models*"]

[tool.pytest.ini_options]
pythonpath = ["."]